Added individual querying and search functionality to MushroomHouseController
Individual querying is based on the MushroomHouse ID. The result will be not found if there is no matching house in the dataset.
It's possible to filter the houses based on a resident's name. By doing so, the result will only contain those houses where there is at least one Smurf, who's name is the searched one.  It's possible to pass the minimum or maximum capacity as a parameter as well. In this case only those houses will be listed whose capacity is in this given interval. Last but not least it's possible to pass the color of the house as well. By doing so, only the houses with this color will be included. It's possible to combine all of these in a request. If there's no house with these parameters, the result will be an empty list.