Added individual querying and search functionality to WorkingPlaceController
Individual querying is based on the WorkingPlace ID. The result will be not found if there is no matching WorkingPlace in the dataset.
There are three basic ways to search for workplaces. It's possible to pass a name parameter, and the result will only contain the workplaces where the name matches the parameter.  You can pass an employee name as well, and then the response will only include the workplaces where there's at least one employee with this name. Last but not least it's also possible to pass a job, and the result will only contain those workplaces where that job is accepted.
I also fixed a performance issue in the queries. The problem was that i used ToLower() every time to compare the texts, but it's not effective in big tables. So with future expansions in mind i rather used the Like Entity Framework function.
There was a logic error in my code too. After parsing the given string parameters to Enums i didn't check if the operation was correct. So when it wasn't, the value that was compared with the Enums in the table was zero, and the condition was true if the Enum in the table had zero value as well, resulting in false positives.