Created a general service interface for the entities
This will serve as a foundation for the actual entity service interfaces such as ISmurfService or IMushroomHouseService etc. This increases expansibility because if there is a new common function needed for all of the services ,there's only one place where this has to be done. This will enforce all of the concrete implementations to adapt to the changes.