Implemented the filter classes to facilitate searching and expansibility
I implemented filter classes because they make the code more readable. Now there's no need to list all of the query parameters in the controller functions. They also help in unifying the service interfaces because the type of the filter has to be added when the concrete entity service is created, eg. ISmurfService : IEntityService<SmurfDto, SmurfFilterDto>.