Implemented SmurfService and changed SmurfController to use it
I implemented the SmurfService class which implements the ISmurfService interface. I moved almost everything from the SmurfController functions to the SmurfService functions and now the SmurfController methods only contain a call to the given SmurfService methods and a return statement. The SmurfController component gets the ISmurfService implementation with dependency injection. For that i had to add the implementation to the services:
builder.Services.AddScoped<ISmurfService, SmurfService>();
I created a constructor for the SmurfService class with the database context as a parameter, so it can access the context for the database calls.
I also rewrote some of the DTO-s so they now use properties instead of fields which - at least for me - increases readability when the DTO has to be created.