Implemented the service responsible for exporting dwarfes' data
I implemented the ExportService class, which is responsible for exporting the  dwarfes' data. It gets the exporting strategies in an IEnumerable through dependency injection. These strategies implement the actual exporting algorithms. Now there's only one strategy, which is TxtExportSmurfsStrategy. This class is responsible for exporting the data into a string which is returned as a byte array.
The export service class is injected into the SmurfsController, which delegates the exporting to this service in case of the the Smurfs/actions/export/{type} and the Smurfs/{id}/actions/export/{type} endpoints. The type here is the type in which the smurf(s) has/have to be exported, eg. txt.