Added DTOs for the already existing utils and fixed some issues
I changed the mapping of the entities to the DTOs in the previous commit so it could be unified in a single function, but i didn't know that select won't work if i call a mapping function inside of it. So i had to change it back to how it was.