Added insert functionality to the services
I added insert functionality to the entity services.  When a post request is sent with the according createdto in the body, every service checks the constraints that are given. For example when the user wants to update a workplace, the service checks if the accepted jobs are compatible with the employees' qualifications